+++ aim/web/ui/src/pages/RunDetail/RunOverviewTab/components/ArtifactsCard/RunOverviewTabArtifactsCard.d.ts
@@ -1,0 +1,4 @@
+export interface IRunOverviewTabArtifactsCardProps {
+  artifacts: { [key: string]: string };
+  isRunInfoLoading: boolean;
+}