+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## 3.19.0
+## 3.19.0 Mar 13, 2024
 
 ### Enhancements:
 - Replace grpc with http/ws as transport for aim tracking server (mihran113)
 - Remove `aim storage upgrade 2to3` command (mihran113)
+- Allow HF callback to initialize run at on_init_end for tracking custom metrics with callback (dushyantbehl)
+- Support artifacts logging and storage in AWS S3 (alberttorosyan)
+- Always set run name when initializing Run in lightning callback (martenlienen) 
 
 ### Fixes
 - Allow the web UI to serve assets symlinked into the static files directory (martenlienen)